FIX:Problems with listing field type photo
I got a problem with the listing field type photo. Listing owners can
upload a picture of a plan in the middle of the listing detail page.
Everything is working fine, when adding a new listing and uploading the
photo.

But there is a problem, if the customer will edit his listing. After
saving the changes, the uploaded image from the photo field disappeared.
Maybe its auto deleted after editing a listing?

Do someone else works with this listing field type and got the same
problems?